+++ packages/generator-superset/generators/plugin-chart/templates/package.erb
@@ -29,8 +29,7 @@
     "@superset-ui/chart": "^0.14.1",
     "@superset-ui/core": "^0.15.0",
     "@superset-ui/chart-controls": "^0.14.0",
-    "@superset-ui/validator": "^0.14.1",
-    "@superset-ui/style": "^0.14.0",
+    "@superset-ui/validator": "^0.14.1"
   },
   "peerDependencies": {
     "react": "^16.13.1"
