+++ packages/superset-ui-core/package.json
@@ -32,6 +32,8 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.1.2",
+    "@emotion/core": "^10.0.28",
+    "@emotion/styled": "^10.0.27",
     "@types/d3-format": "^1.3.0",
     "@types/lodash": "^4.14.149",
     "@types/d3-time": "^1.0.9",
@@ -43,6 +45,7 @@
     "d3-format": "^1.3.2",
     "d3-time": "^1.0.10",
     "d3-time-format": "^2.2.0",
+    "emotion-theming": "^10.0.27",
     "jed": "^1.1.1",
     "lodash": "^4.17.11",
     "fetch-retry": "^4.0.1",
