+++ packages/superset-ui-core/src/index.ts
@@ -26,3 +26,4 @@ export * from './number-format';
 export * from './time-format';
 export * from './dimension';
 export * from './color';
+export * from './style';
