+++ packages/superset-ui-demo/.storybook/themeDecorator.js
@@ -1,6 +1,6 @@
 // themeDecorator.js
 import React from "react"
-import { supersetTheme, ThemeProvider } from '@superset-ui/style';
+import { supersetTheme, ThemeProvider } from '@superset-ui/core';
 
 const ThemeDecorator = storyFn => (
   <ThemeProvider theme={supersetTheme}>{storyFn()}</ThemeProvider>
