+++ packages/superset-ui-demo/storybook/stories/superset-ui-style/ThemeStories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { supersetTheme } from '@superset-ui/style/src/';
+import { supersetTheme } from '@superset-ui/core';
 
 export default {
   title: 'Core Packages|@superset-ui/style',
