+++ plugins/legacy-plugin-chart-calendar/package.json
@@ -37,7 +37,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@superset-ui/validator": "^0.14.0",
     "react": "^16.13.1"
   }
