+++ plugins/legacy-plugin-chart-calendar/src/ReactCalendar.jsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { reactify } from '@superset-ui/chart';
-import styled from '@superset-ui/style';
+import { styled } from '@superset-ui/core';
 import Component from './Calendar';
 
 const ReactComponent = reactify(Component);
