+++ plugins/legacy-plugin-chart-chord/tsconfig.json
@@ -16,9 +16,6 @@
   ],
   "references": [
     {
-      "path": "../../packages/superset-ui-style"
-    },
-    {
       "path": "../../packages/superset-ui-core"
     },
     {
