+++ plugins/legacy-plugin-chart-force-directed/package.json
@@ -31,7 +31,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.3",
     "d3": "^3.5.17",
     "prop-types": "^15.7.2"
   },
