+++ plugins/legacy-plugin-chart-force-directed/src/ReactForceDirected.jsx
@@ -18,7 +18,7 @@
  */
 import React from 'react';
 import { reactify } from '@superset-ui/chart';
-import styled from '@superset-ui/style';
+import { styled } from '@superset-ui/core';
 import PropTypes from 'prop-types';
 import Component from './ForceDirected';
