+++ plugins/legacy-plugin-chart-force-directed/tsconfig.json
@@ -23,9 +23,6 @@
     },
     {
       "path": "../../packages/superset-ui-core"
-    },
-    {
-      "path": "../../packages/superset-ui-style"
     }
   ]
 }
