+++ plugins/legacy-plugin-chart-parallel-coordinates/package.json
@@ -30,7 +30,6 @@
   "dependencies": {
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.3",
     "d3": "^3.5.17",
     "prop-types": "^15.7.2"
   },
