+++ plugins/legacy-plugin-chart-sankey/package.json
@@ -30,7 +30,6 @@
   "dependencies": {
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.3",
     "prop-types": "^15.6.2",
     "d3": "^3.5.17",
     "d3-sankey": "^0.4.2"
