+++ plugins/legacy-plugin-chart-sankey/tsconfig.json
@@ -20,9 +20,6 @@
     },
     {
       "path": "../../packages/superset-ui-core"
-    },
-    {
-      "path": "../../packages/superset-ui-style"
     }
   ]
 }
