+++ plugins/legacy-plugin-chart-time-table/package.json
@@ -32,7 +32,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.3",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@superset-ui/validator": "^0.14.0",
     "@types/d3-scale": "^2.0.2",
     "@types/react": "^16.9.43",
