+++ plugins/legacy-plugin-chart-time-table/src/TimeTable.tsx
@@ -20,8 +20,7 @@ import React from 'react';
 import Mustache from 'mustache';
 import { scaleLinear } from 'd3-scale';
 import { Table, Thead, Th, Tr, Td } from 'reactable-arc';
-import { formatNumber, formatTime } from '@superset-ui/core';
-import styled from '@superset-ui/style';
+import { formatNumber, formatTime, styled } from '@superset-ui/core';
 import moment from 'moment';
 import { InfoTooltipWithTrigger, MetricOption, Metric } from '@superset-ui/chart-controls';
