+++ plugins/legacy-plugin-chart-world-map/package.json
@@ -31,7 +31,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.3",
     "d3": "^3.5.17",
     "d3-array": "^2.4.0",
     "d3-color": "^1.4.1",
