+++ plugins/legacy-plugin-chart-world-map/src/ReactWorldMap.jsx
@@ -18,7 +18,7 @@
  */
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from '@superset-ui/style';
+import { styled } from '@superset-ui/core';
 import { reactify } from '@superset-ui/chart';
 import WorldMap from './WorldMap';
