+++ plugins/legacy-preset-chart-big-number/package.json
@@ -32,7 +32,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@types/d3-color": "^1.2.2",
     "@types/shortid": "^0.0.29",
     "d3-color": "^1.2.3",
