+++ plugins/legacy-preset-chart-nvd3/package.json
@@ -32,7 +32,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.20",
     "@superset-ui/validator": "^0.14.0",
     "d3": "^3.5.17",
     "d3-tip": "^0.9.1",
