+++ plugins/legacy-preset-chart-nvd3/src/ReactNVD3.jsx
@@ -18,7 +18,7 @@
  */
 import React from 'react';
 import { reactify } from '@superset-ui/chart';
-import styled from '@superset-ui/style';
+import { styled } from '@superset-ui/core';
 import PropTypes from 'prop-types';
 import Component from './NVD3Vis';
 import { hideTooltips, removeTooltip } from './utils';
