+++ plugins/legacy-preset-chart-nvd3/tsconfig.json
@@ -25,9 +25,6 @@
       "path": "../../packages/superset-ui-core"
     },
     {
-      "path": "../../packages/superset-ui-style"
-    },
-    {
       "path": "../../packages/superset-ui-validator"
     }
   ]
