+++ plugins/plugin-chart-choropleth-map/package.json
@@ -30,7 +30,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-composition": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@types/d3-geo": "^1.11.1",
     "@types/geojson": "^7946.0.3",
     "@types/react": "^16.9.43",
