+++ plugins/plugin-chart-choropleth-map/src/chart/components.ts
@@ -1,4 +1,4 @@
-import styled, { supersetTheme } from '@superset-ui/style';
+import { styled, supersetTheme } from '@superset-ui/core';
 
 export const PADDING = supersetTheme.gridUnit * 4;
