+++ plugins/plugin-chart-table/package.json
@@ -30,7 +30,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@superset-ui/validator": "^0.14.0",
     "@types/d3-array": "^2.0.0",
     "@types/match-sorter": "^4.0.0",
