+++ plugins/plugin-chart-table/src/Styles.tsx
@@ -1,4 +1,4 @@
-import styled from '@superset-ui/style';
+import { styled } from '@superset-ui/core';
 
 export default styled.div`
   table {
