+++ plugins/plugin-chart-word-cloud/package.json
@@ -30,7 +30,6 @@
   "dependencies": {
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@superset-ui/validator": "^0.14.0",
     "@types/d3-cloud": "^1.2.1",
     "@types/d3-scale": "^2.0.2",
