+++ plugins/plugin-chart-word-cloud/src/chart/WordCloud.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cloudLayout, { Word } from 'd3-cloud';
 import { PlainObject, createEncoderFactory, DeriveEncoding } from 'encodable';
-import { SupersetThemeProps, withTheme } from '@superset-ui/style';
+import { SupersetThemeProps, withTheme } from '@superset-ui/core';
 
 export const ROTATION = {
   flat: () => 0,
