+++ plugins/plugin-chart-word-cloud/tsconfig.json
@@ -22,9 +22,6 @@
       "path": "../../packages/superset-ui-core"
     },
     {
-      "path": "../../packages/superset-ui-style"
-    },
-    {
       "path": "../../packages/superset-ui-validator"
     }
   ]
