+++ temporary-plugins/hold-potentially-deprecate/legacy-plugin-chart-table/package.json
@@ -31,7 +31,6 @@
     "@superset-ui/chart": "^0.14.0",
     "@superset-ui/chart-controls": "^0.14.0",
     "@superset-ui/core": "^0.15.0",
-    "@superset-ui/style": "^0.14.0",
     "@superset-ui/validator": "^0.14.0",
     "@types/datatables.net": "^1.10.18",
     "@types/react-dom": "^16.9.8",
